+++ uniCloud-aliyun/cloudfunctions/router/service/muban_easy.js
@@ -8,7 +8,7 @@ module.exports = {
 	 */
 	main: async (event) => {
 		let { data = {}, userInfo, util, filterResponse, originalParam } = event;
-		let { customUtil, config, pubFun, vk, db, _ } = util;
+		let { customUtil, uniID, config, pubFun, vk, db, _ } = util;
 		let { uid } = data;
 		let res = { code: 0, msg: "" };
 		// 业务逻辑开始-----------------------------------------------------------